+++ SimChA/DataTypes/SubClone.cs
@@ -8,19 +8,19 @@ public class SubClone
     public int CloneId { get; }
     public int ParentId { get; }
     public double DivisionRate { get; }
-    public uint NumberDrivers { get; }
-    private List<(uint Alive, uint Dead)> Cells { get; }
+    public int NumberDrivers { get; }
+    private List<(long Alive, long Dead)> Cells { get; }
     public double ToDivide { get; set; }
     public double ToDie { get; set; }
     
     // public Karyotype Karyotype { get; }
 
-    public uint AliveCount => Cells.Last().Alive;
-    public uint DeadCount => Cells.Last().Dead;
-    public uint TotalCount => AliveCount + DeadCount;
+    public long AliveCount => Cells.Last().Alive;
+    public long DeadCount => Cells.Last().Dead;
+    public long TotalCount => AliveCount + DeadCount;
     public int LastGen => FirstGen + Cells.Count;
     
-    public SubClone(int cloneId, int parentId, int generation, double divisionRate, uint numberDrivers = 1, uint popSize = 1) 
+    public SubClone(int cloneId, int parentId, int generation, double divisionRate, int numberDrivers = 1, uint popSize = 1) 
     {
         CloneId = cloneId;
         ParentId = parentId;
@@ -28,22 +28,22 @@ public class SubClone
         // Karyotype = new Karyotype(karyotype);
         DivisionRate = Math.Clamp(divisionRate, 0, 1);
         FirstGen = generation;
-        Cells = new List<(uint, uint)> { (popSize, 0) };
+        Cells = new List<(long, long)> { (popSize, 0) };
         ToDivide = 0;
         ToDie = 0;
     }
     
-    public SubClone CreateChild(int newId, int generation, double divRateChange, uint numberDrivers)
+    public SubClone CreateChild(int newId, int generation, double divRateChange, int numberDrivers)
         => new(newId, CloneId, generation, divRateChange, numberDrivers);
     
     public override string ToString() 
         => $"ID:{CloneId}, Parent:{ParentId}, Alive: {AliveCount}, " +
            $"Dead: {DeadCount}, Drivers: {NumberDrivers}, DivisionRate: {DivisionRate}";
 
-    public uint AliveAtGen(int gen)
+    public long AliveAtGen(int gen)
         => gen >= FirstGen && gen < LastGen ? Cells[gen - FirstGen].Alive : 0;
 
-    public uint TotalAtGen(int gen)
+    public long TotalAtGen(int gen)
     {
         return gen >= FirstGen && gen < LastGen ? Cells[gen - FirstGen].Alive + Cells[gen - FirstGen].Dead : 0;
     }
